database/node_database.py
"""
Node Database
PostgreSQL database integration for node-level data persistence
Handles local node state, worker tracking, and task history
"""

import psycopg2
import psycopg2.pool
from psycopg2.extras import RealDictCursor
import logging
from typing import Optional, List, Dict, Any
from datetime import datetime
import json
import structlog

logger = structlog.get_logger(__name__)


class NodeDatabase:
    """
    PostgreSQL database interface for node-level operations
    Manages local node state, worker registry, and task tracking
    """
    
    def __init__(self, connection_config: Dict[str, Any]):
        """Initialize database connection"""
        self.connection_config = connection_config
        self.connection_pool = None
        
        # Initialize connection pool
        self._create_connection_pool()
        
        # Initialize database schema
        self._initialize_schema()
        
        logger.info("NodeDatabase initialized")
    
    def _create_connection_pool(self):
        """Create PostgreSQL connection pool"""
        # TODO: Implement connection pool creation
        # 1. Create threaded connection pool
        # 2. Set pool parameters
        # 3. Test connection
        pass
    
    def _initialize_schema(self):
        """Initialize database schema if needed"""
        # TODO: Implement schema initialization
        # 1. Check if tables exist
        # 2. Create tables if needed
        # 3. Run migrations
        pass
    
    def get_connection(self):
        """Get database connection from pool"""
        # TODO: Get connection from pool
        pass
    
    def return_connection(self, conn):
        """Return connection to pool"""
        # TODO: Return connection to pool
        pass
    
    def execute_query(self, query: str, params: Optional[tuple] = None) -> List[Dict[str, Any]]:
        """Execute query and return results"""
        # TODO: Implement query execution
        # 1. Get connection
        # 2. Execute query
        # 3. Handle results
        # 4. Return connection
        return []
    
    # Node management methods
    def update_node_status(self, node_id: str, status: str, details: Optional[Dict[str, Any]] = None):
        """Update node status in database"""
        # TODO: Update node status
        pass
    
    def record_node_heartbeat(self, node_id: str):
        """Record node heartbeat"""
        # TODO: Update last heartbeat timestamp
        pass
    
    # Worker management methods
    def register_worker(self, worker_info: Dict[str, Any]) -> bool:
        """Register a worker in database"""
        # TODO: Insert worker record
        pass
    
    def update_worker_status(self, worker_id: str, status: str, details: Optional[Dict[str, Any]] = None):
        """Update worker status"""
        # TODO: Update worker status
        pass
    
    def get_active_workers(self) -> List[Dict[str, Any]]:
        """Get list of active workers"""
        # TODO: Query active workers
        return []
    
    # Task management methods
    def create_task_record(self, task_info: Dict[str, Any]) -> str:
        """Create task record and return task ID"""
        # TODO: Insert task record
        return "task-id"
    
    def update_task_status(self, task_id: str, status: str, result: Optional[Any] = None):
        """Update task status and result"""
        # TODO: Update task record
        pass
    
    def get_task_history(self, limit: int = 100) -> List[Dict[str, Any]]:
        """Get task execution history"""
        # TODO: Query task history
        return []
    
    # Resource monitoring methods
    def record_resource_usage(self, resource_data: Dict[str, Any]):
        """Record resource usage metrics"""
        # TODO: Insert resource metrics
        pass
    
    def get_resource_history(self, hours: int = 24) -> List[Dict[str, Any]]:
        """Get resource usage history"""
        # TODO: Query resource history
        return []
    
    def cleanup_old_records(self, days: int = 7):
        """Clean up old records"""
        # TODO: Delete old records
        # 1. Delete old task records
        # 2. Delete old resource metrics
        # 3. Keep recent data
        pass
    
    def close(self):
        """Close database connections"""
        # TODO: Close connection pool
        logger.info("NodeDatabase connections closed")
